requirements.txt
discord.py
python-dotenv
openai
pytest
pytest-asyncio